requirements.txt
amqp==5.0.6
asgiref==3.3.1
astroid==2.3.3
billiard==3.6.4.0
boto3==1.17.48
botocore==1.20.48
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
Django==3.1.7
django-appconf==1.0.4
django-crontab==0.7.1
django-imagekit==4.0.2
django-ipware==3.0.2
django-phonenumber-field==5.0.0
django-solo==1.1.5
idna==2.10
isort==4.3.21
jmespath==0.10.0
kombu==5.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
metaflow==2.2.8
phonenumbers==8.12.20
pilkit==2.0
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pylint==2.4.4
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
wrapt==1.11.2
